change oracle size
as `maxSize = 2**16 - 1`, and `id < maxSize`, the Oracle will never have more than 65_535 samples

+++ src/interfaces/ILBPair.sol
@@ -41,7 +41,7 @@ interface ILBPair {
         uint16 oracleSize;
         uint16 oracleActiveSize;
         uint40 oracleLastTimestamp;
-        uint24 oracleId;
+        uint16 oracleId;
         FeeHelper.FeesDistribution feesX;
         FeeHelper.FeesDistribution feesY;
     }
